requirements.txt
streamlit>=1.28.0
litellm>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 
langchain>=0.0.208
python-dotenv>=1.0.0
pdfplumber
numpy>=1.24.0
